pyproject.toml
[tool.poetry]
name = "pennywise"
version = "0.1.0"
description = ""
authors = ["Samuel Adekoya <damiloladrizzy@yahoo.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
boto3 = "^1.35.92"
moto = "^5.0.26"
pytest = "^8.3.4"
google-generativeai = "^0.8.3"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
